Implement battle system and game state management
- Added BattleResult enum and BattleSystem struct to handle battle logic.
- Introduced PlayerAction enum for player actions in the game.
- Created GamePhase enum to manage game phases.
- Developed GameState struct to track player and game state.
- Enhanced UI with status display and action input handling.
- Implemented helper functions for random name generation and number formatting.
- Updated region difficulty type from u8 to u32 for better scalability.

+++ src/world/region.rs
@@ -6,7 +6,7 @@ pub struct Region {
     pub id: u32,
     pub name: String,
     pub description: String,
-    pub difficulty: u8,
+    pub difficulty: u32,
     pub resources: Vec<String>,
     pub controlled_by: Option<u32>, // Faction ID
 }
